dom/base/test/test_bug1399605.html

<!DOCTYPE HTML>
<html>
<head>
  <meta charset="utf-8">
  <title>Test for Bug 1399605</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=1399605">Mozilla Bug 1399605</a>
<p id="display"></p>
<div id="content">
  <div id="target"></div>
</div>
<pre id="test">
<script type="application/javascript">

  let observer = new IntersectionObserver(function() {
    ok(true, "we are still observing");
    SimpleTest.finish();
  });
  observer.observe(document.getElementById('target'));
  observer.unobserve(document.getElementById('content'));
  SimpleTest.waitForExplicitFinish();

</script>
</pre>
<div id="log">
</div>
</body>
</html>
